Conflicts:
	arch/arm64/boot/dts/amlogic/Makefile
	arch/arm64/boot/dts/amlogic/g12a_s905x2_u215.dts
	arch/arm64/boot/dts/amlogic/g12b_a311d_w400.dts
	arch/arm64/boot/dts/amlogic/g12b_a311d_w400_linux.dts
	arch/arm64/boot/dts/amlogic/mesonsm1.dtsi
	arch/arm64/boot/dts/amlogic/sm1_s905c1_ac232.dts
	arch/arm64/boot/dts/amlogic/t3x_t968d4_bc303.dts
	drivers/Kconfig
	drivers/Makefile
	drivers/drm/meson_hdmi.c
	drivers/drm/vpu-hw/meson_vpu_osd_mif.c
	drivers/media/vin/tvin/vdin/vdin_v4l2_if.h
	drivers/media/vout/hdmitx21/hdmi_tx_main.c
	drivers/media/vout/hdmitx21/hw/hw_clk.c
	drivers/media/vout/hdmitx_common/hdmitx_mode.c
	sound/soc/amlogic/auge/earc.c
	sound/soc/amlogic/auge/pdm.c

+++ include/uapi/amlogic/msync.h
@@ -24,7 +24,7 @@ struct pts_tri {
 	uint32_t wall_clock;
 	uint32_t pts;
 	uint32_t delay;
-	uint32_t mono_ts;
+	uint64_t mono_ts;
 };
 
 struct pcr_pair {
